Reorganize argument passing and docs (#98)
* Reorganize argument passing and docs

* Reorganize high level API

* Fix mocks

* Refactor CLI

* Fix parsing of CGNC

* Fix forcing

* Update logging

* Update test_extract.py

* Enable passing of strict option to xref db build

* Enable --no-strict CLI option for database build

* Fix import

* Make version keyword non optional

Even if none is passed, something always should be passed

* Add caching and loading of definitions

also add direct lookup for synonyms

* Cleanup code

* Add more curations for typos in Gene Ontology

* Fix interpro tree parser

* Fix mocks

* Fix definition parsing and allow specification of strictness

+++ src/pyobo/aws.py
@@ -12,12 +12,14 @@ import humanize
 from more_click import verbose_option
 from tabulate import tabulate
 
-from pyobo.constants import RAW_DIRECTORY
-from pyobo.extract import (
-    get_id_name_mapping, get_id_synonyms_mapping, get_id_to_alts, get_properties_df, get_relations_df, get_xrefs_df,
-    iter_cached_obo,
+from pyobo import (
+    get_id_name_mapping, get_id_synonyms_mapping, get_id_to_alts, get_properties_df, get_relations_df,
+    get_xrefs_df,
 )
+from pyobo.api.utils import get_version
+from pyobo.constants import RAW_DIRECTORY
 from pyobo.path_utils import prefix_cache_join
+from pyobo.registries import iter_cached_obo
 
 __all__ = [
     'download_artifacts',
@@ -84,42 +86,54 @@ def upload_artifacts_for_prefix(*, prefix: str, bucket: str, s3_client=None):
 
     logger.info('[%s] getting id->name mapping', prefix)
     get_id_name_mapping(prefix)
-    id_name_path = prefix_cache_join(prefix, 'names.tsv')
+    id_name_path = prefix_cache_join(prefix, 'names.tsv', version=get_version(prefix))
+    if not id_name_path.exists():
+        raise FileNotFoundError
     id_name_key = os.path.join(prefix, 'cache', 'names.tsv')
     logger.info('[%s] uploading id->name mapping', prefix)
     upload_file(path=id_name_path, bucket=bucket, key=id_name_key, s3_client=s3_client)
 
     logger.info('[%s] getting id->synonyms mapping', prefix)
     get_id_synonyms_mapping(prefix)
-    id_synonyms_path = prefix_cache_join(prefix, 'synonyms.tsv')
+    id_synonyms_path = prefix_cache_join(prefix, 'synonyms.tsv', version=get_version(prefix))
+    if not id_synonyms_path.exists():
+        raise FileNotFoundError
     id_synonyms_key = os.path.join(prefix, 'cache', 'synonyms.tsv')
     logger.info('[%s] uploading id->synonyms mapping', prefix)
     upload_file(path=id_synonyms_path, bucket=bucket, key=id_synonyms_key, s3_client=s3_client)
 
     logger.info('[%s] getting xrefs', prefix)
     get_xrefs_df(prefix)
-    xrefs_path = prefix_cache_join(prefix, 'xrefs.tsv')
+    xrefs_path = prefix_cache_join(prefix, 'xrefs.tsv', version=get_version(prefix))
+    if not xrefs_path.exists():
+        raise FileNotFoundError
     xrefs_key = os.path.join(prefix, 'cache', 'xrefs.tsv')
     logger.info('[%s] uploading xrefs', prefix)
     upload_file(path=xrefs_path, bucket=bucket, key=xrefs_key, s3_client=s3_client)
 
     logger.info('[%s] getting relations', prefix)
     get_relations_df(prefix)
-    relations_path = prefix_cache_join(prefix, 'relations.tsv')
+    relations_path = prefix_cache_join(prefix, 'relations.tsv', version=get_version(prefix))
+    if not relations_path.exists():
+        raise FileNotFoundError
     relations_key = os.path.join(prefix, 'cache', 'relations.tsv')
     logger.info('[%s] uploading relations', prefix)
     upload_file(path=relations_path, bucket=bucket, key=relations_key, s3_client=s3_client)
 
     logger.info('[%s] getting properties', prefix)
     get_properties_df(prefix)
-    properties_path = prefix_cache_join(prefix, 'properties.tsv')
+    properties_path = prefix_cache_join(prefix, 'properties.tsv', version=get_version(prefix))
+    if not properties_path.exists():
+        raise FileNotFoundError
     properties_key = os.path.join(prefix, 'cache', 'properties.tsv')
     logger.info('[%s] uploading properties', prefix)
     upload_file(path=properties_path, bucket=bucket, key=properties_key, s3_client=s3_client)
 
     logger.info('[%s] getting alternative identifiers', prefix)
     get_id_to_alts(prefix)
-    alts_path = prefix_cache_join(prefix, 'alt_ids.tsv')
+    alts_path = prefix_cache_join(prefix, 'alt_ids.tsv', version=get_version(prefix))
+    if not alts_path.exists():
+        raise FileNotFoundError
     alts_key = os.path.join(prefix, 'cache', 'alt_ids.tsv')
     logger.info('[%s] uploading alternative identifiers', prefix)
     upload_file(path=alts_path, bucket=bucket, key=alts_key)
